add delete confirmation

+++ application/views/pages/admin/creative/images.php
@@ -4,11 +4,11 @@
 		<?php
 		// Display the edit form (from controller)
 		if ($ad_image) { 
-			echo '<div>Ad Image | <a href="creative_image_delete?creative_id=' .$creative_id. '&image_id='.$ad_id.'"> Delete</a> | <br />'.HTML::image('/media/ad_images/'.$ad_image).'<br /><br /></div>';
+			echo '<div>Ad Image | <a href="creative_image_delete?creative_id=' .$creative_id. '&image_id='.$ad_id.'" onclick="return confirm(\'Do you really want to delete?\')";> Delete</a> <br />'.HTML::image('/media/ad_images/'.$ad_image).'<br /><br /></div>';
 			echo PHP_EOL;
 		}
 		if ($detail_image) {
-			echo '<div>Detail Image | <a href="creative_image_delete?creative_id=' .$creative_id. '&image_id='.$detail_id.'">Delete</a> | <br /> '.HTML::image('/media/ad_images/'.$detail_image).'</div>';
+			echo '<div>Detail Image | <a href="creative_image_delete?creative_id=' .$creative_id. '&image_id='.$detail_id.'" onclick="return confirm(\'Do you really want to delete?\')";>Delete</a> <br /> '.HTML::image('/media/ad_images/'.$detail_image).'</div>';
 			echo PHP_EOL;
 		}
 		echo $form;
